requirements.txt
Django==1.7.1
Pillow==2.5.3
Unidecode==0.04.18
argparse==1.2.1
django-forms-bootstrap==3.0.0
django-uuslug==1.0.2
python-slugify==1.1.2
wsgiref==0.1.2
